+++ Sisapco/src/main/java/co/com/sisapco/repository/AlmacenamientoGoogleRepository.java
@@ -8,7 +8,7 @@ import co.com.sisapco.entity.Actividades;
 import co.com.sisapco.entity.AlmacenamientoGoogle;
 
 @Repository
-public interface AlmacenamientoGoogleRepository extends CrudRepository<AlmacenamientoGoogle, Long>{
+public interface AlmacenamientoGoogleRepository extends CrudRepository<AlmacenamientoGoogle, Integer>{
 
 	@Query("SELECT u FROM AlmacenamientoGoogle u WHERE u.copNit = ?1 AND u.almaCarpeta = ?2")
 	public AlmacenamientoGoogle findByAlmacenamientoGoogleNit(int codNit, String carpeta);
